requirements.txt
Django == 3.1.4

djangorestframework == 3.13.0

psycopg2>=2.7.5,<2.8.0

Pillow == 8.4.0
